+++ src/components/forms/Form.ts
@@ -2,8 +2,7 @@
 // Imports
 //
 
-import { Child, DE, ElementAttributes, ElementAttributesSchema } from "@donutteam/document-builder";
-import { z } from "zod";
+import { Child, DE, ElementAttributes } from "@donutteam/document-builder";
 
 import { HiddenInput } from "./HiddenInput.js";
 
@@ -11,44 +10,35 @@ import { HiddenInput } from "./HiddenInput.js";
 // Component
 //
 
-export const ProtectionOptionsSchema = z.union(
-	[
-		z.object(
-			{
-				type: z.literal("none"),
-			}),
+export type FormProtectionOptions =
+{
+	type: "none";
+} |
+{
+	type: "recaptcha";
+	siteKey: string;
+};
 
-		z.object(
-			{
-				type: z.literal("recaptcha"),
-				siteKey: z.string(),
-			}),
-	]);
+export type FormOptions =
+{
+	className?: string;
 
-export type ProtectionOptions = z.infer<typeof ProtectionOptionsSchema>;
+	method?: "get" | "GET" | "post" | "POST" | "dialog" | "DIALOG";
+	action?: string;
 
-export const FormOptionsSchema = z.object(
-	{
-		className: z.string().optional(),
+	autoComplete?: boolean; 
+	encodingType?: "application/x-www-form-urlencoded" | "multipart/form-data" | "text/plain";
 
-		method: z.enum([ "get", "GET", "post", "POST", "dialog", "DIALOG" ]).optional(),
-		action: z.string().optional(),
+	manuallyInitialize?: boolean;
+	maxFileSize?: number;
+	noticeContainerSelector?: string;
 
-		autoComplete: z.boolean().optional(),
-		encodingType: z.enum([ "application/x-www-form-urlencoded", "multipart/form-data", "text/plain" ]).optional(),
+	protection?: FormProtectionOptions;
 
-		manuallyInitialize: z.boolean().optional(),
-		maxFileSize: z.number().optional(),
-		noticeContainerSelector: z.string().optional(),
+	hiddenInputs?: Record<string, string>;
 
-		protection: ProtectionOptionsSchema.optional(),
-
-		hiddenInputs: z.record(z.string()).optional(),
-
-		extraAttributes: ElementAttributesSchema.optional(),
-	});
-
-export type FormOptions = z.infer<typeof FormOptionsSchema>;
+	extraAttributes?: ElementAttributes;
+};
 
 export function Form(options: FormOptions, children: Child)
 {